Reconcile ambiguous backup creation during atomic upload
